+++ NEWS.md
@@ -1,8 +1,33 @@
+# feasts 0.1.2
+
+## Improvements
+
+* `var_tiled_var()` no longer includes partial tile windows in the computation.
+* Added residual acf features to `feat_stl()`.
+* Performance improvements.
+
+## Breaking changes
+
+* Decompositions are now treated as models. 
+  To access the decomposed values, you will now have to use `components()`.
+  For example, `tourism %>% STL(Trips)` is now `tourism %>% model(STL(Trips)) %>% components()`.
+  This change allows for more flexible decomposition specifications, and better interfaces for decomposition modelling.
+
+## Bug fixes
+
+* Fixed bug with `feat_spectral()` not showing results.
+* Fix warning in `ACF()`, `PACF()` and `CCF()` for tidyr change.
+* `gg_tsdisplay()` will no longer fail on non-seasonal data with missing values. The last plot will instead show a PACF in this case (#76)
+* Better handling of perfect fits in `stat_arch_lm()` (#85)
+
 # feasts 0.1.1
 
 ## Improvements
 
 * Better naming of seasonal columns in STL decomposition when seasonal period is specified.
+
+## Bug fixes
+
 * Fixes issues with running tests on unsupported systems.
 
 # feasts 0.1.0